+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.9.2] - 2022-09-22
+### Changed
+
+- Email verification endpoints will now clear the session if called by a deleted/unknown user
+
+
+## [0.9.1] - 2022-09-20
+
+### Adds:
+
+- Adds Dashboard recipe
+
+
 ## [0.9.0] - 2022-09-14
 
 ### Added